+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This changelog is initialized in release 1.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2022-09-26
+
+### Added
+* SourceRowInterface::getGroupByName
+* SourceRowInterface::getFirstDataValueByPlaceholder
+
 ## [2.0.0] - 2022-09-14
 
 ### Added
@@ -22,6 +28,7 @@ This changelog is initialized in release 1.0.0
 ### Added
 * Possibility to limit generated selectors
 
-[Unreleased]: https://github.com/wimski/html-data-extractor/compare/2.0.0...master
+[Unreleased]: https://github.com/wimski/html-data-extractor/compare/2.1.0...master
+[2.1.0]: https://github.com/wimski/html-data-extractor/compare/2.0.0...2.1.0
 [2.0.0]: https://github.com/wimski/html-data-extractor/compare/1.1.0...2.0.0
 [1.1.0]: https://github.com/wimski/html-data-extractor/compare/1.0.0...1.1.0